+++ src/netspresso_trainer/losses/detection/retinanet.py
@@ -4,9 +4,9 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch import Tensor
+from torchvision.models.detection._utils import BoxCoder, Matcher
 from torchvision.ops import boxes as box_ops
 
-from ...models.heads.detection.experimental.detection._utils import BoxCoder, Matcher
 from ..common import SigmoidFocalLoss