+++ src/netspresso_trainer/models/registry.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 from .backbones import cspdarknet, efficientformer, mixnet, mobilenetv3, mobilevit, resnet, segformer, vit
 from .full import pidnet
 from .heads.classification import fc
-from .heads.detection import faster_rcnn, retinanet_head, yolox_head
+from .heads.detection import retinanet_head, yolox_head
 from .heads.segmentation import all_mlp_decoder
 from .necks import fpn, pafpn
 
@@ -34,7 +34,6 @@ MODEL_HEAD_DICT: Dict[str, Callable[..., nn.Module]] = {
         'all_mlp_decoder': all_mlp_decoder,
     },
     'detection': {
-        'faster_rcnn': faster_rcnn,
         'yolox_head': yolox_head,
         'retinanet_head': retinanet_head,
     },